feat: add img desing and corrections directory AB#2
