+++ src/app/(main)/(private)/context/page.tsx
@@ -1,10 +1,12 @@
 "use client"
-import ContextPush from "@/ui/molecules/steps/ContextPush";
+import ContextPush from "@/ui/molecules/contextpush/ContextPush";
 import DataCollection from "@/ui/organisms/dataCollection/DataCollection";
 
 export default function ContextView(){
     return (
+        <div className="containergeneral">
 
-       <ContextPush/>
+            <ContextPush/>
+        </div>
     )
 }
