+++ src/app/styles.scss
@@ -1,0 +1,9 @@
+.containergeneral {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    
+}
