+++ src/ui/molecules/contextpush/ContextPush.tsx
@@ -1,23 +1,24 @@
 import React, { useState } from 'react';
 import { useUserStore } from '../../../app/core/application/global-state/store';
-import styles from './StepOne.module.scss';
-
+import styles from './ContextPush.module.scss';
+import Image from 'next/image';
 
 const ContextPush: React.FC = () => {
 
 
   return (
-    <div className={styles.stepContainer}>
+    <div className={styles.contextpush}>
       <div className={styles.header}>
-        <img
-          src="../../../public/joseui.png"
-          alt="joseiu"
-          className={styles.icon}
-        />
         <h2 className={styles.title}>wait a moment, we are preparing your routine</h2>
       </div>
       <div className={styles.optionsGrid}>
-
+        <Image
+          src="/img/joseui.png"
+          alt="Jose Iu Smart"
+          className={styles.icon}
+          width={400} // Ajusta el ancho
+          height={400} // Ajusta el alto
+        />
 
       </div>
       <button
