+++ src/ui/molecules/progressBar/ProgressBar.module.scss
